+++ src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/worker/common/Runner.kt
@@ -46,6 +46,7 @@ import com.tencent.devops.worker.common.logger.LoggerService
 import com.tencent.devops.worker.common.service.ProcessService
 import com.tencent.devops.worker.common.task.TaskDaemon
 import com.tencent.devops.worker.common.task.TaskFactory
+import com.tencent.devops.worker.common.utils.KillBuildProcessTree
 import org.slf4j.LoggerFactory
 import java.io.File
 import kotlin.system.exitProcess
@@ -59,7 +60,8 @@ object Runner {
             logger.info("Start the worker ...")
             // 启动成功了，报告process我已经启动了
             val buildVariables = ProcessService.setStarted()
-
+            // 为进程加上ShutdownHook事件
+            KillBuildProcessTree.addKillProcessTreeHook(buildVariables.projectId, buildVariables.buildId, buildVariables.vmSeqId)
             // 启动日志服务
             LoggerService.start()
             val variables = buildVariables.variablesWithType
