+++ src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/worker/common/utils/KillBuildProcessTree.kt
@@ -24,7 +24,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.agent.utils
+package com.tencent.devops.worker.common.utils
 
 import com.tencent.devops.common.api.enums.OSType
 import com.tencent.devops.worker.common.env.AgentEnv
@@ -58,6 +58,20 @@ object KillBuildProcessTree {
             -1
         } finally {
             reader?.close()
+        }
+    }
+
+    fun addKillProcessTreeHook(projectId: String, buildId: String, vmSeqId: String) {
+        try {
+            Runtime.getRuntime().addShutdownHook(object : Thread() {
+                override fun run() {
+                    logger.info("start kill process tree")
+                    val killedProcessIds = killProcessTree(projectId, buildId, vmSeqId)
+                    logger.info("kill process tree done, ${killedProcessIds.size} process(s) killed, pid(s): $killedProcessIds")
+                }
+            })
+        } catch (t: Throwable) {
+            logger.warn("Fail to add shutdown hook", t)
         }
     }
